requirements.txt
certifi==2017.4.17
chardet==3.0.4
crate==0.20.0
idna==2.5
inflection==0.3.1
python-intercom==3.1.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1